test: extract parse-hanson-string tests into separate files
