+++ modules/hanson-format/__tests__/parse-hanson-string/filter-expression.test.js
@@ -1,0 +1,8 @@
+import {expect} from 'chai'
+import {parse} from '../../parse-hanson-string'
+
+describe('FilterExpression', () => {
+	it('', () => {
+
+	})
+})
